# frontend/packages/api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"tags":[{"name":"Wallet API V1","description":"API for wallet balances and transfer history"},{"name":"Device API V1","description":"Retired device registration compatibility endpoint"},{"name":"Governance API V1","description":"Authority membership and BIP state"}],"paths":{"/api/core/v1/wallet/submit-tx":{"post":{"tags":["Wallet API V1"],"summary":"Submit transaction","description":"Submit a transaction","operationId":"submitTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TxSubmitDtoV1"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MempoolResult"}}}}}}},"/api/core/v1/node-webhook/handle":{"post":{"tags":["node-webhook-api-v-1"],"operationId":"receiveWebhook","parameters":[{"name":"X-Webhook-Timestamp","in":"header","required":true,"schema":{"type":"string"}},{"name":"X-Webhook-Signature","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/core/v1/device/register":{"post":{"tags":["Device API V1"],"summary":"Acknowledge legacy device registration","description":"Rolling-retirement endpoint; persistence remains configurable until old backend replicas drain","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRegistrationRequestDtoV1"}}},"required":true},"responses":{"200":{"description":"Legacy registration acknowledged","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDtoV1"}}}}}}},"/api/core/v1/wallet/transfers":{"get":{"tags":["Wallet API V1"],"summary":"Get transfer history","description":"Get unified transfer history (pending first, then confirmed)","operationId":"getTransfers","parameters":[{"name":"addresses","in":"query","description":"1–100 wallet addresses; an empty list is rejected","required":true,"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}},{"name":"tokenAddresses","in":"query","description":"At most 100 token addresses (optional; omitted means all tokens)","required":false,"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}},{"name":"transferType","in":"query","description":"Transfer type (optional, null for all types)","required":false,"schema":{"type":"string","enum":["TRANSFER","BLOCK_FEES","BLOCK_REWARD","MINT","BURN"]}},{"name":"pageNumber","in":"query","description":"Page number (0-indexed; pageNumber × pageSize must not exceed 100000)","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","description":"Page size, 1–100","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UnifiedTransferPageDtoV1"}}}}}}},"/api/core/v1/wallet/transaction-status":{"get":{"tags":["Wallet API V1"],"summary":"Observe transaction status","description":"Observe a previously signed transaction without submitting or replaying it","operationId":"getTransactionStatus","parameters":[{"name":"hash","in":"query","description":"Canonical lowercase transaction hash","required":true,"schema":{"type":"string","pattern":"^0x[0-9a-f]{64}$"}},{"name":"sender","in":"query","description":"Transaction sender","required":true,"schema":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"}},{"name":"nonce","in":"query","description":"Canonical uint256 decimal nonce","required":true,"schema":{"type":"string","maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionStatusDtoV1"}}}}}}},"/api/core/v1/wallet/tokens":{"get":{"tags":["Wallet API V1"],"summary":"Get tokens","description":"Get paginated list of available tokens","operationId":"getTokens","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TokenDtoV1"}}}}}}}},"/api/core/v1/wallet/token":{"get":{"tags":["Wallet API V1"],"summary":"Get token by address","description":"Get token details by contract address","operationId":"getTokenByAddress","parameters":[{"name":"address","in":"query","description":"Token contract address","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDtoV1"}}}}}}},"/api/core/v1/wallet/next-nonce":{"get":{"tags":["Wallet API V1"],"summary":"Get next nonce","description":"Get next nonce for a given address","operationId":"getNextNonce","parameters":[{"name":"address","in":"query","description":"Address","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Canonical non-negative decimal next nonce","content":{"*/*":{"schema":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"}}}}}}},"/api/core/v1/wallet/mempool-recommended-fees":{"get":{"tags":["Wallet API V1"],"summary":"Get mempool recommended fees","description":"Get mempool recommended fees","operationId":"getMempoolRecommendedFees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MempoolRecommendedFeesDtoV1"}}}}}}},"/api/core/v1/wallet/balances":{"get":{"tags":["Wallet API V1"],"summary":"Get wallet balances","description":"Get total, locked, spendable and available balances for 1–100 addresses; large results must be narrowed","operationId":"getBalances","parameters":[{"name":"addresses","in":"query","description":"1–100 wallet addresses; an empty list is rejected","required":true,"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}},{"name":"tokenAddresses","in":"query","description":"At most 100 token addresses (optional; omitted means all tokens)","required":false,"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WalletBalanceDtoV1"}}}}}}}},"/api/core/v1/governance/options":{"get":{"tags":["Governance API V1"],"summary":"Get current entities available to BIP creation forms","operationId":"getOptions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GovernanceOptionsDtoV1"}}}}}}},"/api/core/v1/governance/bips":{"get":{"tags":["Governance API V1"],"summary":"Get the BIP overview","operationId":"getBips","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0,"maximum":100000,"minimum":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20,"maximum":100,"minimum":1}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BipPageDtoV1"}}}}}}},"/api/core/v1/governance/bip":{"get":{"tags":["Governance API V1"],"summary":"Get a BIP by its creation transaction hash","operationId":"getBip","parameters":[{"name":"hash","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BipDtoV1"}}}}}}},"/api/core/v1/governance/authority-status":{"get":{"tags":["Governance API V1"],"summary":"Check whether an address is a current authority","operationId":"getAuthorityStatus","parameters":[{"name":"address","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorityStatusDtoV1"}}}}}}}},"components":{"schemas":{"TxSubmitDtoV1":{"type":"object","description":"Transaction details","properties":{"hexData":{"type":"string","maxLength":200002,"minLength":0,"pattern":"^0x(?:[0-9a-fA-F]{2})+$"}},"required":["hexData"]},"MempoolResult":{"type":"object","properties":{"status":{"type":["string","null"],"enum":["SUCCESS","QUEUED","STALE","REJECTED_FEE","REJECTED_RBF","REJECTED_STATE","REJECTED_DUPLICATE","REJECTED_NONCE_TOO_FAR_FUTURE","REJECTED_MEMPOOL_FULL","REJECTED_OTHER"]},"message":{"type":["string","null"]}}},"DeviceRegistrationRequestDtoV1":{"type":"object","properties":{"clientIdentifier":{"type":"string","format":"uuid"},"platform":{"type":"string","maxLength":20,"minLength":0},"fcmToken":{"type":"string","maxLength":512,"minLength":0},"appVersion":{"type":"string","maxLength":20,"minLength":0}},"required":["clientIdentifier"]},"DeviceDtoV1":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"clientIdentifier":{"type":"string","format":"uuid"},"platform":{"type":"string"},"fcmToken":{"type":"string"},"appVersion":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"lastSeenAt":{"type":"string","format":"date-time"}}},"UnifiedTransferDtoV1":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","CONFIRMED"]},"txHash":{"type":"string"},"transferType":{"type":"string","enum":["TRANSFER","BLOCK_FEES","BLOCK_REWARD","MINT","BURN"]},"from":{"type":"string"},"to":{"type":"string"},"tokenAddress":{"type":"string"},"amount":{"type":"string"},"fee":{"type":"string"},"nonce":{"type":"string","description":"Canonical non-negative decimal transaction nonce","pattern":"^(0|[1-9][0-9]*)$"},"message":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"blockHeight":{"type":"string","description":"Canonical non-negative decimal block height","pattern":"^(0|[1-9][0-9]*)$"},"blockHash":{"type":"string"},"confirmations":{"type":"string","description":"Canonical non-negative decimal confirmation count","pattern":"^(0|[1-9][0-9]*)$"}}},"UnifiedTransferPageDtoV1":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/UnifiedTransferDtoV1"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"string","description":"Canonical non-negative decimal total transfer count","pattern":"^(0|[1-9][0-9]*)$"},"totalPages":{"type":"integer","format":"int32"},"pendingCount":{"type":"string","description":"Canonical non-negative decimal pending transfer count","pattern":"^(0|[1-9][0-9]*)$"},"confirmedCount":{"type":"string","description":"Canonical non-negative decimal confirmed transfer count","pattern":"^(0|[1-9][0-9]*)$"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"TransactionStatusDtoV1":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","CONFIRMING","CONFIRMED","ABSENT_REUSABLE","CONSUMED_SUPERSEDED","BLOCKED_UNKNOWN"]},"hash":{"type":"string","pattern":"^0x[0-9a-f]{64}$"},"sender":{"type":"string","pattern":"^0x[0-9a-f]{40}$"},"nonce":{"type":"string","maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"},"nextNonce":{"type":["string","null"],"maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"},"confirmations":{"type":["string","null"],"maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"},"requiredConfirmations":{"type":"string","maxLength":78,"minLength":0,"pattern":"^[1-9][0-9]*$"}},"required":["confirmations","hash","nextNonce","nonce","requiredConfirmations","sender","status"]},"TokenDtoV1":{"type":"object","properties":{"address":{"type":"string"},"name":{"type":"string"},"smallestUnitName":{"type":"string"},"numberOfDecimals":{"type":"integer","format":"int32"},"websiteUrl":{"type":"string"},"logoUrl":{"type":"string"},"maxSupply":{"type":"string"},"userBurnable":{"type":"boolean"},"originTxHash":{"type":"string"},"totalSupply":{"type":"string"}}},"MempoolRecommendedFeesDtoV1":{"type":"object","properties":{"slow":{"$ref":"#/components/schemas/MempoolRecommendedFeesLevelDtoV1"},"standard":{"$ref":"#/components/schemas/MempoolRecommendedFeesLevelDtoV1"},"fast":{"$ref":"#/components/schemas/MempoolRecommendedFeesLevelDtoV1"},"mempoolSize":{"type":"string","description":"Canonical non-negative decimal mempool transaction count","pattern":"^(0|[1-9][0-9]*)$"}},"required":["fast","mempoolSize","slow","standard"]},"MempoolRecommendedFeesLevelDtoV1":{"type":"object","properties":{"baseFee":{"type":"string","maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"},"feePerByte":{"type":"string","maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"},"totalForAverageTx":{"type":"string","maxLength":78,"minLength":0,"pattern":"^(0|[1-9][0-9]*)$"}},"required":["baseFee","feePerByte","totalForAverageTx"]},"WalletBalanceDtoV1":{"type":"object","properties":{"address":{"type":"string"},"tokenAddress":{"type":"string"},"balance":{"type":"string","description":"Available amount after pending outgoing amounts and native fees; used by Send/MAX"},"updatedAtBlockHeight":{"type":"string","description":"Canonical non-negative decimal block height","pattern":"^(0|[1-9][0-9]*)$"},"updatedAtTimestamp":{"type":"string","format":"date-time"},"totalBalance":{"type":"string","description":"Full confirmed amount including locked mining rewards"},"lockedMiningReward":{"type":"string","description":"Confirmed mining rewards that cannot yet be spent"},"spendableBalance":{"type":"string","description":"Confirmed unlocked amount before pending outgoing reservations"}},"required":["totalBalance"]},"AddressAliasOptionDtoV1":{"type":"object","properties":{"alias":{"type":"string"},"address":{"type":"string"}},"required":["address","alias"]},"GovernanceOptionsDtoV1":{"type":"object","properties":{"authorities":{"type":"array","items":{"type":"string"}},"addressAliases":{"type":"array","items":{"$ref":"#/components/schemas/AddressAliasOptionDtoV1"}},"validators":{"type":"array","items":{"$ref":"#/components/schemas/ValidatorOptionDtoV1"}},"addressAliasesTruncated":{"type":"boolean"}},"required":["addressAliases","addressAliasesTruncated","authorities","validators"]},"ValidatorOptionDtoV1":{"type":"object","properties":{"address":{"type":"string"},"miningLimitMode":{"type":"string"},"maxMiningShareBps":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"}},"required":["address","maxMiningShareBps","miningLimitMode"]},"BipDtoV1":{"type":"object","properties":{"bipHash":{"type":"string"},"status":{"type":"string"},"actionExecuted":{"type":"boolean"},"type":{"type":"string"},"numberOfRequiredVotes":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"},"approvers":{"type":"array","items":{"type":"string"}},"disapprovers":{"type":"array","items":{"type":"string"}},"executedAtTimestamp":{"type":"string","format":"date-time"},"expirationTimestamp":{"type":"string","format":"date-time"},"createdAtBlockHeight":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"},"createdAtTimestamp":{"type":"string","format":"date-time"},"updatedAtBlockHeight":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"},"updatedAtTimestamp":{"type":"string","format":"date-time"},"updatedByTxHash":{"type":"string"},"metadata":{"$ref":"#/components/schemas/BipMetadataDtoV1"}},"required":["actionExecuted","approvers","bipHash","createdAtBlockHeight","createdAtTimestamp","disapprovers","expirationTimestamp","metadata","numberOfRequiredVotes","status","type","updatedAtBlockHeight","updatedAtTimestamp","updatedByTxHash"]},"BipMetadataDtoV1":{"type":"object","properties":{"txVersion":{"type":"string"},"derivedTokenAddress":{"type":"string"},"txPayload":{"type":"object","additionalProperties":{}}},"required":["txPayload","txVersion"]},"BipPageDtoV1":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/BipDtoV1"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"string","pattern":"^(0|[1-9][0-9]*)$"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}},"required":["content","first","last","pageNumber","pageSize","totalElements","totalPages"]},"AuthorityStatusDtoV1":{"type":"object","properties":{"address":{"type":"string"},"authority":{"type":"boolean"}},"required":["address","authority"]}},"securitySchemes":{"BasicAuth":{"type":"http","description":"Basic Auth for Wallet Admins","scheme":"basic"}}}}